Changes to frap simulation
* added a frap end time as a user input
* removed the frapresult matrix and calculation, FRAP is simulated within the C++ now
* changed the pyx file to allow for cases where user does not provide a probe, ie no fluorsecent probe. The C++ will still run the simulation with intensity set to zero if the user wants a tasep with no fluorescence
* removed old model maker from the rsnapsim as its now defunct
* defaulted poi probevec and probeloc to np.int32
* colors are now set by the probe provided / protein given to the simulation, not outside the simulation
* updated warnings for the user with the simulation

+++ rsnapsim/poi.py
@@ -164,7 +164,7 @@ class poi():
         for i in range(len(list(self.tag_epitopes))):
             pv[i, [self.tag_epitopes[list(self.tag_epitopes.keys())[i]]]] = 1
         pv = np.cumsum(pv, axis=1)
-        return pv
+        return pv.astype(np.int32)
 
     @property
     def probe_loc(self):
@@ -177,7 +177,7 @@ class poi():
             binary vector of epitope locations over Ncolor x L of the transcript.
 
         '''
-        pv = np.zeros((len(list(self.tag_epitopes)), self.total_length))
+        pv = np.zeros((len(list(self.tag_epitopes)), self.total_length), dtype = np.int32)
         for i in range(len(list(self.tag_epitopes))):
             pv[i, [self.tag_epitopes[list(self.tag_epitopes.keys())[i]]]] = 1
         return pv
